telegram-channel-acl-bot/requirements.txt
pyrogram==2.0.106
python-dotenv==1.0.1
https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip